# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tickets API",
        "description": "API para gerenciamento de funcionários.",
        "contact": {
            "email": "jhonattan.curtarelli@outlook.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/tickets": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Busca todos os tickets",
                "operationId": "63410598a093cab144b5580f73c75839",
                "responses": {
                    "200": {
                        "description": "Lista de tickets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ticket"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Cria um novo ticket",
                "operationId": "64c186f3265e66b70e42d675d1821d84",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "employee_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "employee_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "situation": {
                                        "type": "string",
                                        "enum": [
                                            "A",
                                            "I"
                                        ],
                                        "example": "A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erros de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Dados Inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tickets/{id}": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Busca o ticket pelo código",
                "operationId": "f276270bce513c39d80ba96831f613c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Código do ticket",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do ticket",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ticket não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Atualiza o ticket",
                "operationId": "e6d92b0d6a2df06e8a397a2ff206f495",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do ticket",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "situation": {
                                        "type": "string",
                                        "enum": [
                                            "A",
                                            "I"
                                        ],
                                        "example": "A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ticket atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket não encontrado"
                    },
                    "422": {
                        "description": "Erros de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Dados Inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Deleta um ticket",
                "operationId": "0bb37b0cac18d58ae5df81b261a1e65d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do ticket",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ticket deletado com sucesso"
                    },
                    "404": {
                        "description": "Ticket não encontrado"
                    },
                    "400": {
                        "description": "Erro ao deletar o ticket",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao deletar o ticket"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Ticket": {
                "title": "Tickets",
                "description": "Tickets",
                "required": [
                    "id",
                    "employee_id",
                    "quantity",
                    "situation"
                ],
                "properties": {
                    "id": {
                        "description": "ID do ticket",
                        "type": "integer",
                        "example": 1
                    },
                    "employee_id": {
                        "description": "ID do funcionário associado ao ticket",
                        "type": "integer",
                        "example": 1
                    },
                    "quantity": {
                        "description": "Quantidade de tickets",
                        "type": "integer",
                        "example": 5
                    },
                    "situation": {
                        "description": "Situação do ticket",
                        "type": "string",
                        "enum": [
                            "A",
                            "I"
                        ],
                        "example": "A"
                    },
                    "created_at": {
                        "description": "Data de criação do ticket",
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-09-01T00:00:00"
                    },
                    "updated_at": {
                        "description": "Data de atualização do ticket",
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-09-01T00:00:00"
                    },
                    "deleted_at": {
                        "description": "Data de exclusão do ticket",
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-09-01T00:00:00"
                    }
                },
                "type": "object"
            }
        }
    },
...
}